Added sentry logging. Fixed issues with quotas. Update to latest Go version

+++ server/pkg/quota/quota.go
@@ -1,7 +1,7 @@
 package quota
 
-const BandwidthQuotaInMiB int32 = 1024 * 1000 // 1000 Gib
+const BandwidthQuotaInMiB int64 = 1024 * 1000 // 1000 Gib
 
-func BytesAsMib(numBytes int64) int32 {
-	return int32(numBytes / 1048576)
+func BytesAsMib(numBytes int64) int64 {
+	return numBytes / 1048576
 }
